fakebtc/requirements.txt
fonttools==4.44.3
fpdf2==2.7.7
pyfiglet==1.0.2
requests==2.31.0
pytest==7.4.3
